+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Added` for new features
 - dist: subscriber table, therefore subscription input box is working now
 - dist: featureFlag `newletter_input_box` can turn off both the (un)subscribe email input box and the POST value processing
+- L10n (Localisation) class with loadLocalisation and translate methods common both for admin and MyCMS
 
 ### `Changed` for changes in existing functionality
 - App class handles the request dispatching instead of spagetti code in index.php